pythonDjangoDevelopEnv/requirements.txt
Django==3.2.5
numpy
DateTime
requests
django-rest-framework
django-cors-headers
Pillow